Codegen quality improvements: performance, naming, and Infer

+++ tests/basics/record_match/record_match.cpp
@@ -1,8 +1,8 @@
 #include "record_match.h"
 
-unsigned int RecordMatch::sum(const RecordMatch::MyRec &r) {
-  unsigned int n1 = r.f1;
-  unsigned int n2 = r.f2;
-  unsigned int n3 = r.f3;
+uint64_t RecordMatch::sum(const RecordMatch::MyRec &r) {
+  uint64_t n1 = r.f1;
+  uint64_t n2 = r.f2;
+  uint64_t n3 = r.f3;
   return ((n1 + n2) + n3);
 }
