+++ tests/regression/empty_match/empty_match.cpp
@@ -1,5 +1,5 @@
 #include "empty_match.h"
 
-unsigned int EmptyMatch::from_empty(const EmptyMatch::empty &_x0) {
-  return absurd<unsigned int>(_x0);
+uint64_t EmptyMatch::from_empty(const EmptyMatch::empty &_x0) {
+  return absurd<uint64_t>(_x0);
 }
