+++ tests/regression/list_closure_escape/list_closure_escape.cpp
@@ -7,10 +7,8 @@ ListClosureEscape::fn_list
 ListClosureEscape::build_fns(ListClosureEscape::tree t1,
                              ListClosureEscape::tree t2) {
   return fn_list::fcons(
-      [=](unsigned int _x0) mutable -> unsigned int {
-        return t1.sum_values(_x0);
-      },
-      fn_list::fcons([=](unsigned int _x0) mutable
-                         -> unsigned int { return t2.sum_values(_x0); },
-                     fn_list::fnil()));
+      [=](uint64_t _x0) mutable -> uint64_t { return t1.sum_values(_x0); },
+      fn_list::fcons(
+          [=](uint64_t _x0) mutable -> uint64_t { return t2.sum_values(_x0); },
+          fn_list::fnil()));
 }
