+++ tests/regression/move_capture_reuse/move_capture_reuse.cpp
@@ -1,8 +1,8 @@
 #include "move_capture_reuse.h"
 
-List<List<unsigned int>>
-MoveCaptureReuse::prefix_each(const List<unsigned int> &prefix,
-                              const List<List<unsigned int>> &xss) {
-  return xss.template map<List<unsigned int>>(
-      [=](const List<unsigned int> &xs) mutable { return prefix.app(xs); });
+List<List<uint64_t>>
+MoveCaptureReuse::prefix_each(const List<uint64_t> &prefix,
+                              const List<List<uint64_t>> &xss) {
+  return xss.template map<List<uint64_t>>(
+      [=](const List<uint64_t> &xs) mutable { return prefix.app(xs); });
 }
