+++ tests/regression/move_safety/move_safety.cpp
@@ -5,7 +5,6 @@
 /// the Box will hold a dangling reference after the
 /// function returns.
 MoveSafety::fn_box MoveSafety::make_box(MoveSafety::tree t) {
-  return fn_box::box([=](unsigned int _x0) mutable -> unsigned int {
-    return t.sum_values(_x0);
-  });
+  return fn_box::box(
+      [=](uint64_t _x0) mutable -> uint64_t { return t.sum_values(_x0); });
 }
