+++ tests/regression/nested_record_update_qual/nested_record_update_qual.cpp
@@ -2,6 +2,6 @@
 
 NestedRecordUpdateQual::Shadow
 NestedRecordUpdateQual::bump(const NestedRecordUpdateQual::Shadow &x) {
-  unsigned int n = x.value;
+  uint64_t n = x.value;
   return Shadow{(n + 1)};
 }
