+++ tests/wip/sep_ext_nullary_module_alias/SepExtNullaryModuleAlias.h
@@ -2,9 +2,6 @@
 #define INCLUDED_SEPEXTNULLARYMODULEALIAS
 
 #include <concepts>
-#include <memory>
-#include <optional>
-#include <type_traits>
 
 namespace SepExtNullaryModuleAlias {
 
@@ -24,10 +21,10 @@ concept Config = requires {
   requires HasVal<typename M::V>;
   requires(
       requires {
-        { M::default_val } -> std::convertible_to<unsigned int>;
+        { M::default_val } -> std::convertible_to<uint64_t>;
       } ||
       requires {
-        { M::default_val() } -> std::convertible_to<unsigned int>;
+        { M::default_val() } -> std::convertible_to<uint64_t>;
       });
 };
 
@@ -39,8 +36,8 @@ template <Config C> struct Worker {
     return v;
   }
 
-  static const unsigned int &get_default() {
-    static const unsigned int v = C::default_val();
+  static const uint64_t &get_default() {
+    static const uint64_t v = C::default_val();
     return v;
   }
 };
