+++ tests/wip/sep_ext_nullary_nested_access/SepExtNullaryNestedAccess.h
@@ -2,9 +2,6 @@
 #define INCLUDED_SEPEXTNULLARYNESTEDACCESS
 
 #include <concepts>
-#include <memory>
-#include <optional>
-#include <type_traits>
 
 namespace SepExtNullaryNestedAccess {
 
@@ -12,10 +9,10 @@ template <typename M>
 concept Inner = requires {
   requires(
       requires {
-        { M::val } -> std::convertible_to<unsigned int>;
+        { M::val } -> std::convertible_to<uint64_t>;
       } ||
       requires {
-        { M::val() } -> std::convertible_to<unsigned int>;
+        { M::val() } -> std::convertible_to<uint64_t>;
       });
 };
 template <typename M>
@@ -23,26 +20,26 @@ concept Outer = requires {
   requires Inner<typename M::I>;
   requires(
       requires {
-        { M::name } -> std::convertible_to<unsigned int>;
+        { M::name } -> std::convertible_to<uint64_t>;
       } ||
       requires {
-        { M::name() } -> std::convertible_to<unsigned int>;
+        { M::name() } -> std::convertible_to<uint64_t>;
       });
 };
 
 template <Outer O> struct Worker {
-  static const unsigned int &get_inner_val() {
-    static const unsigned int v = O::I::val();
+  static const uint64_t &get_inner_val() {
+    static const uint64_t v = O::I::val();
     return v;
   }
 
-  static const unsigned int &get_name() {
-    static const unsigned int v = O::name();
+  static const uint64_t &get_name() {
+    static const uint64_t v = O::name();
     return v;
   }
 
-  static const unsigned int &sum() {
-    static const unsigned int v = (O::I::val() + O::name());
+  static const uint64_t &sum() {
+    static const uint64_t v = (O::I::val() + O::name());
     return v;
   }
 };
